[bp/r24720][FIX] mail notifications : fix the link sent in the notification email which gives the possibility to unsubscribe the user watch.     BTW, fix the remove_user_watch_by_id function in order to not allow anybody to remove every notification.     This was broken since r11994 and was ticket 2895 in dev tracker.     Thks to nyloth and sept_7 for the help
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/proposals@24722 b456876b-0849-0410-b77d-98878d47e9d5

+++ 3.x/lib/tikilib.php
@@ -250,10 +250,30 @@ class TikiLib extends TikiDB {
 		return true;
 	}
 
+	/**
+	 * get_user_notification: returns the owner (user) related to a watchId 
+	 * 
+	 * @param mixed $id watchId 
+	 * @access public
+	 * @return the user login related to the watchId
+	 */
+	function get_user_notification($id) {
+
+		$query = "select `user` from `tiki_user_watches` where `watchId`=?";
+		return $this->getOne($query, array($id));
+
+	}
 	/*shared*/
 	function remove_user_watch_by_id($id) {
-		$query = "delete from `tiki_user_watches` where `watchId`=?";
-		$this->query($query,array($id));
+		global $tiki_p_admin_notifications, $user;
+		if ( $tiki_p_admin_notifications === 'y' or $user === $this->get_user_notification($id) ) {
+			$query = "delete from `tiki_user_watches` where `watchId`=?";
+			$this->query($query, array($id));
+
+			return true;
+		}
+
+		return false;
 	}
 
 	function remove_group_watch_by_id($id) {
